% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pfpj.R
\name{arruma_sa}
\alias{arruma_sa}
\title{as funções dessa parte do pacote são responsáveis por criar as colunas
que serão efetivamente analisadas no relatório}
\usage{
arruma_sa(string_vec)
}
\description{
as funções dessa parte do pacote são responsáveis por criar as colunas
que serão efetivamente analisadas no relatório
}
