% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pfpj.R
\name{market_segments}
\alias{market_segments}
\title{Creates column of companies segment}
\usage{
market_segments(names_column)
}
\arguments{
\item{names_column}{a string vector of company names}
}
\value{
the string vector of company segments
}
\description{
Receives a vector of company names and return a vector os segments.
}
