% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sg-day.R
\name{sg_day}
\alias{sg_day}
\title{Downloads and parses one day}
\usage{
sg_day(date = Sys.Date() - 1, path = ".", verbose = TRUE,
  folder_prefix = "sg_")
}
\arguments{
\item{date}{date to download. Default is today minus one.}

\item{path}{path to directory where to create the folder.}

\item{verbose}{print messages. Default is to print messages.}

\item{folder_prefix}{folder prefix to create. Default is \code{sg_}.}
}
\description{
Uses \code{esaj} to download and parse all files from cjsg and cposg from
TJSP for one day
}
